Move `reactivateEveryTab` into `@/utils`
